backend/db.py
import mysql.connector

def get_db_connection():
    return mysql.connector.connect(
        host="localhost",
        user="root",
        password="Motog42",
        database="customer_segmentation"
    )
