backend/requirements.txt
absl-py==2.1.0
alembic==1.16.4
altair==5.3.0
annotated-types==0.7.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
autoviz==0.1.905
Babel==2.13.0
backcall==0.2.0
bcrypt==5.0.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.8.2
bokeh==3.6.3
branca==0.7.2
bs4==0.0.2
cachetools==5.3.3
catboost==1.2.8
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
colorcet==3.1.0
colorlog==6.9.0
comm==0.1.4
contourpy==1.3.1
cycler==0.11.0
dacite==1.8.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.4
django-embed-video==1.4.9
dnspython==2.7.0
easy-thumbnails==2.8.5
email_validator==2.2.0
emoji==2.14.1
emojis==0.7.0
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==2.0.0
fastapi==0.115.12
fastjsonschema==2.18.1
Flask==3.0.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.12.23
folium==0.17.0
fonttools==4.37.1
fqdn==1.5.1
fsspec==2025.2.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==25.0.3
h11==0.14.0
h5py==3.12.1
holoviews==1.20.1
html5lib==1.1
htmlmin==0.1.12
httplib2==0.31.1
hvplot==0.11.2
idna==3.4
ImageHash==4.3.1
imbalanced-learn==0.13.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.16.1
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==3.8.0
kiwisolver==1.4.4
libclang==18.1.1
lightgbm==4.6.0
linkify-it-py==2.0.3
llvmlite==0.43.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.10.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.0.2
ml_dtypes==0.5.1
mlxtend==0.23.1
multimethod==1.11.2
mysql-connector==2.2.9
mysql-connector-python==9.2.0
mysqlclient==2.2.7
namex==0.0.8
nbclient==0.8.0
nbconvert==7.9.1
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.3
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.60.0
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.3
opt_einsum==3.4.0
optree==0.14.0
optuna==4.4.0
outcome==1.2.0
overrides==7.4.0
packaging==21.3
pandas==2.2.3
pandas-dq==1.29
pandas-profiling==3.6.6
pandocfilters==1.5.0
panel==1.6.1
paperclip==2.7.2
param==2.2.0
parso==0.8.3
patsy==0.5.6
phik==0.12.4
pickleshare==0.7.5
Pillow==9.2.0
pip-review==1.3.0
platformdirs==3.11.0
plotly==5.24.1
pluggy==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.27.0
protobuf==5.27.1
psutil==5.9.5
pure-eval==0.2.2
pyamg==5.2.1
pyarrow==16.1.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybind11==2.13.6
pycparser==2.21
pydantic==2.7.4
pydantic-settings==2.5.2
pydantic_core==2.18.4
pydeck==0.9.1
pygame==2.6.1
Pygments==2.16.1
PyJWT==2.10.1
pyparsing==3.0.9
pypng==0.20220715.0
PyQRCode==1.2.1
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.27
pytz==2023.3
pyviz_comms==3.0.4
PyWavelets==1.6.0
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.30.2
regex==2024.5.15
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.3
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
selenium==4.11.2
Send2Trash==1.8.2
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.38
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.2
streamlit==1.36.0
streamlit_folium==0.20.1
tenacity==8.4.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.17.1
textblob==0.18.0.post0
threadpoolctl==3.5.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.3.3
tqdm==4.66.4
traitlets==5.11.2
trio==0.22.2
trio-websocket==0.10.3
typeguard==4.3.0
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2023.3
uc-micro-py==1.0.3
unicorn==2.1.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.0.4
uvicorn==0.34.3
visions==0.7.6
watchdog==4.0.1
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==3.0.3
widgetsnbextension==4.0.9
wordcloud==1.9.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
xgboost==1.6.2
xlrd==2.0.1
XlsxWriter==3.2.0
xyzservices==2024.6.0
ydata-profiling==4.8.3
